TrashCan complete. Until tomorrow...

+++ src/utils/trashCan.ts
@@ -2,7 +2,7 @@ import { CopyCustomComp, CopyNativeEl, OrigCustomComp, AppInterface, Parent } fr
 import { Originals, Copies } from './interfaces';
 import { isCopyCustomComp } from './parser';
 
-export const trashCan = (compToDelete: CopyNativeEl | CopyCustomComp, copyOriginals: Originals, copyCopies: Copies) => {
+export const trashCan = (compToDelete: CopyNativeEl | CopyCustomComp, copyOriginals: Originals, copyCopies: Copies, deleteFromCanvas=false) => {
   // For any CopyNativeEl, we need to delete:
   // (1) the reference in parent's children array
   // (2) itself in the copies context
@@ -25,7 +25,7 @@ export const trashCan = (compToDelete: CopyNativeEl | CopyCustomComp, copyOrigin
     } else if (compToDeleteParent.origin === 'copies') {
       const parentChildToDelete = copyCopies[compToDeleteParent.key] as CopyNativeEl;
       const originalParentElement = copyOriginals[parentChildToDelete.pointer] as OrigCustomComp;
-      const parentChildren: string[] = isCopyCustomComp(parentChildToDelete) ? originalParentElement.children : parentChildToDelete.children;
+			const parentChildren: string[] = isCopyCustomComp(parentChildToDelete) ? originalParentElement.children : parentChildToDelete.children;
       const parentChildIdx = parentChildren.indexOf(compToDelete.name);
       parentChildren.splice(parentChildIdx, 1);
     }
@@ -40,37 +40,39 @@ export const trashCan = (compToDelete: CopyNativeEl | CopyCustomComp, copyOrigin
       return;
     }
     
-    // make copy of children array since splicing elements from it while looping over it causes errors 
-    const copyOfCompToDeleteChildren = [...compToDeleteChildren];
-    for (const child of copyOfCompToDeleteChildren) {
-      // if the child of compToDelete is a CopyCustomComp
-      if (isCopyCustomComp(copyCopies[child])) {
-        // delete the copy reference of the child in the parent's children array (from originals context)
-        if (copyCopies[child].parent.origin === 'original') {
-          const parentOfCopy = copyOriginals[copyCopies[child].parent.key] as OrigCustomComp;
-          const parentChildren = parentOfCopy.children;
-          const parentChildIdx = parentChildren.indexOf(child);
-          parentChildren.splice(parentChildIdx, 1);
-        } else if (copyCopies[child].parent.origin === 'copies') {
-          const parentOfCopy = copyCopies[copyCopies[child].parent.key] as CopyNativeEl;
-          const parentChildren = parentOfCopy.children;
-          const parentChildIdx = parentChildren.indexOf(child);
-          parentChildren.splice(parentChildIdx, 1);
+    if (!deleteFromCanvas) {
+      // make copy of children array since splicing elements from it while looping over it causes errors 
+      const copyOfCompToDeleteChildren = [...compToDeleteChildren];
+      for (const child of copyOfCompToDeleteChildren) {
+        // if the child of compToDelete is a CopyCustomComp
+        if (isCopyCustomComp(copyCopies[child])) {
+          // delete the copy reference of the child in the parent's children array (from originals context)
+          if (copyCopies[child].parent.origin === 'original') {
+            const parentOfCopy = copyOriginals[copyCopies[child].parent.key] as OrigCustomComp;
+            const parentChildren = parentOfCopy.children;
+            const parentChildIdx = parentChildren.indexOf(child);
+            parentChildren.splice(parentChildIdx, 1);
+          } else if (copyCopies[child].parent.origin === 'copies') {
+            const parentOfCopy = copyCopies[copyCopies[child].parent.key] as CopyNativeEl;
+            const parentChildren = parentOfCopy.children;
+            const parentChildIdx = parentChildren.indexOf(child);
+            parentChildren.splice(parentChildIdx, 1);
+          }
+
+          // child will be a copy of OrigCustomComp, meaning that child appears in the copies array of its pointer
+          // delete the copy reference of the child in the original's copies array (from originals context)
+          const originalElement = copyOriginals[copyCopies[child].pointer] as OrigCustomComp;
+          const allCopiesInOriginals = originalElement.copies;
+          allCopiesInOriginals.splice(allCopiesInOriginals.indexOf(copyCopies[child].name), 1);
+
+          // delete the child copy object from the copies context
+          delete copyCopies[child];
+        } 
+        // else if child is NOT a CopyCustomComp and the child has not yet been deleted from copies context
+        else if (copyCopies[child]) { 
+          // recursively delete the child in copies context and all of its children and copies
+          deleteCompInCopies(copyCopies[child]);
         }
-
-        // child will be a copy of OrigCustomComp, meaning that child appears in the copies array of its pointer
-        // delete the copy reference of the child in the original's copies array (from originals context)
-        const originalElement = copyOriginals[copyCopies[child].pointer] as OrigCustomComp;
-        const allCopiesInOriginals = originalElement.copies;
-        allCopiesInOriginals.splice(allCopiesInOriginals.indexOf(copyCopies[child].name), 1);
-
-        // delete the child copy object from the copies context
-        delete copyCopies[child];
-      } 
-      // else if child is NOT a CopyCustomComp and the child has not yet been deleted from copies context
-      else if (copyCopies[child]) { 
-        // recursively delete the child in copies context and all of its children and copies
-        deleteCompInCopies(copyCopies[child]);
       }
     }
